Preparing for PHP 8.3 (#279)
* Preparing for PHP 8.3

* temporarily using dev-master for mime dep so 8.3 tests can run

* fixing some deprecations and warnings

* versions

+++ tests/Zend/Mail/MaildirFolderTest.php
@@ -114,7 +114,7 @@ class Zend_Mail_MaildirFolderTest extends PHPUnit\Framework\TestCase
                 closedir($dh);
                 rmdir($this->_tmpdir . $dir . '/' . $subdir);
             }
-            if ($dir != '.') {
+            if ($dir != '.' && is_dir($this->_tmpdir . $dir)) {
                 rmdir($this->_tmpdir . $dir);
             }
         }
